+++ packages/image_picker/image_picker_platform_interface/lib/src/platform_interface/image_picker_platform.dart
@@ -168,6 +168,18 @@ abstract class ImagePickerPlatform extends PlatformInterface {
     throw UnimplementedError('pickVideo() has not been implemented.');
   }
 
+  /// Returns an arbitrary [PickedFile].
+  ///
+  /// The `allowedExtensions` argument controls what files may be selected, for example:
+  /// \['pdf', 'doc'] to allow only to pick .pdf or .doc files.
+  ///
+  /// When `allowedExtensions` is null/empty, the plugin lets the user pick any file.
+  Future<PickedFile> pickArbitraryFile({
+    List<String> allowedExtensions = const [],
+  }) {
+    throw UnimplementedError('pickArbitraryFile() has not been implemented.');
+  }
+
   /// Retrieve the lost [PickedFile] file when [pickImage] or [pickVideo] failed because the MainActivity is destroyed. (Android only)
   ///
   /// Image or video can be lost if the MainActivity is destroyed. And there is no guarantee that the MainActivity is always alive.