+++ packages/image_picker/image_picker_platform_interface/test/picked_file_io_test.dart
@@ -12,6 +12,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:image_picker_platform_interface/image_picker_platform_interface.dart';
 
 final String expectedStringContents = 'Hello, world!';
+final int expectedSize = expectedStringContents.length;
 final Uint8List bytes = utf8.encode(expectedStringContents);
 final File textFile = File('./assets/hello.txt');
 final String textFilePath = textFile.path;
@@ -20,9 +21,18 @@ void main() {
   group('Create with an objectUrl', () {
     final pickedFile = PickedFile(textFilePath);
 
+    test('Can get length', () async {
+      expect(await pickedFile.length(), equals(expectedSize));
+    });
+
+    test('Extracts the filename', () async {
+      expect(pickedFile.name, equals('hello.txt'));
+    });
+
     test('Can be read as a string', () async {
       expect(await pickedFile.readAsString(), equals(expectedStringContents));
     });
+
     test('Can be read as bytes', () async {
       expect(await pickedFile.readAsBytes(), equals(bytes));
     });