+++ packages/image_picker/image_picker_platform_interface/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0
+
+* Add `pickArbitraryFile` method to the platform interface. This allows users to pick any file type, not just images or videos.
+* Add `name` and `length()` methods to PickedFile.
+
 ## 1.1.0
 
 * Introduce PickedFile type for the new API.